refactor: Remove Script.serialize to extension
